+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2.0
+
+* TODO: Atualização para retorno_sucesso_ou_erro_package 2.0.2.
+* TODO: Usecase e repositorio substituidos por RetornoResultadoPresenter.
+* TODO: Novas features carregar imagens da galeria e de link.
+
 ## 1.1.0
 
 * TODO: Ajustes para flutter 2.0.0.